feat(core,durability,fsfs,fusion,index,rerank): expand host adapter, rerank pipeline, and cross-crate infrastructure
Broad infrastructure expansion across seven crates with 670+ new lines
covering host environment adaptation, rerank pipeline maturation,
durability hardening, and query planning refinements.

Core (~270 lines):
- host_adapter.rs: Major expansion (+208 lines) of host environment
  detection, capability probing, and platform abstraction layer
- collectors.rs: Add 43 lines of new result collector implementations
- commit_replay.rs: Expand commit replay with 16 lines of new
  recovery hooks

Durability (~65 lines):
- config.rs: Expand durability config with new WAL/checkpoint tuning
- fsvi_protector.rs: Refine FSVI protection logic
- tantivy_wrapper.rs: Add 35 lines of Tantivy engine lifecycle hooks

FSFS (~70 lines):
- runtime.rs: Expand runtime with 41 lines of new orchestration logic
- interaction_primitives.rs, lifecycle.rs, query_planning.rs: Refinements

Fusion (~70 lines):
- circuit_breaker.rs: Expand circuit breaker with backoff strategies
- searcher.rs: Expand search coordinator with result merging
- Cargo.toml: Add new dependency

Index (~45 lines):
- mrl.rs, quantization.rs, two_tier.rs: Expand MRL, quantization,
  and two-tier index implementations

Rerank (~110 lines):
- pipeline.rs: Major expansion of rerank pipeline with scoring
  normalization and stage composition (+89 lines)
- lib.rs: Expand rerank module exports (+21 lines)

Storage (+20 lines):
- index_metadata.rs, staleness.rs: Expand metadata tracking

Docs: packaging/release/install contract, telemetry taxonomy,
cross-epic telemetry adapter updates. Beads sync.

+++ crates/frankensearch-durability/src/fsvi_protector.rs
@@ -141,9 +141,10 @@ impl FsviProtector {
         }
 
         let repair_size = fs::metadata(&sidecar_path).map_or(0, |metadata| metadata.len());
-        #[allow(clippy::cast_precision_loss)]
+        // Compute in f64 first to avoid double precision loss from u64→f32.
+        #[allow(clippy::cast_precision_loss, clippy::cast_possible_truncation)]
         let overhead_ratio = if source_size > 0 {
-            repair_size as f32 / source_size as f32
+            (repair_size as f64 / source_size as f64) as f32
         } else {
             0.0
         };
